Bevy 0.14 migration (#54)
Migrates to Bevy 0.14. Bevy 0.14 has significant changes to the
animation system, but the most significant for our purposes is that
usage of `EntityPath` throughout the system is replaced by an opaque
`AnimationTargetId`, which is essentially a wrapper around `Uuid`.

Since we relied on the transparent structure of `EntityPath` to find the
parent of a bone in certain nodes (rotation node, inverse kinematics
node), this change forced us to add a `Skeleton` asset type, that upon
load will store the entity hierarchy of an animated scene and map
`AnimationTargetId`s to entity paths. The hierarchy will be created
automatically, but you need to manually create the skeleton asset and
point it to a Gltf scene to use as reference (see updated examples).