2D blend space node (#78)
This PR adds a couple of things:
* A 2D blend space node in the style of Godot's
[AnimationNodeBlendSpace2D](https://docs.godotengine.org/en/stable/classes/class_animationnodeblendspace2d.html).
Inputs to this node are mapped to a 2D plane and triangulated. At
runtime, a `Vec2` parameter is passed in, and the output pose is the
linear combination of the poses mapped to the vertices of the triangle
that the input vector is contained in. If the input vector is outside
all triangles, it is projected to the closest one.

You may note that there is no such thing as a "linear combination of
quaternions". Instead we treat them as a Vec4, compute the linear
combination of the vectors and re-normalize the quaternions. This
usually works reasonably well (see [Gramkow
2001](https://doi.org/10.1023/A:1011129215388)), is independent of the
order of the quaternions and is reasonably fast. The main issue I've
found with this is that, while quaternions `q` and `-q` may represent
the same rotation, when converted to a vector they are quite different
and mixing in some "negated" quaternions can result in nonsensical
output animations. I'll need to do more research on if/how we should
solve this.

* Some debugging tools for animation graphs in the editor. This wasn't
really in the original plan for the PR, but I found myself needing
better tools to debug issues with the blend space node. In particular,
you can now:
  * Play/pause/step frame by frame the active animated scene preview.
* Inspect the node outputs of each node in the graph, including
rendering a preview of a pose output. When combined with frame stepping,
this was quite useful to me.

* Scene and pose previews can now be rotated and zoomed in/out with the
mouse.