Attaching "hitbox" colliders to bones (#97)
It's common to attach colliders to animated bones to act as "hitboxes"
for projectiles; either to determine which hit animation to play or to
handle damage different for each body part.

This PR adds an UI for doing this in the editor. There's also a new
optional asset type `SkeletonColliders` that is used to add colliders to
a skeleton. The UX is not perfect, but is usable for now and has some
quality of life features like configurable symmetry support (you can see
in the picture below that the "transparent" colliders are replicated by
the symmetry configuration, whereas the ones drawn with full opacity are
manually added):

![collider editor
screenshot](https://github.com/user-attachments/assets/68eed757-b057-4d05-8f8f-3203c38af31b)