fix: avoid race condition when spawning animated scenes with colliders (#101)
Unclear why this was triggering a race condition, as the `AnimatedScene`
asset should have a dependency on all subassets. This will fix it in the
short term while I don't have a more "proper" solution.

Perhaps I need to manually check the "recursive dependency load state"
using the asset server?