Prepare for 0.8 release (#110)
I'm in the process of cleaning up a lot of tech debt in this project,
but in the process of doing that there will be some breaking changes to
the serialized data format. To make migration as simple as possible, I'm
cutting a release before these changes come into effect and I'll cut
another one after; I'm also preparing a small migration guide that I'll
post together with the breaking change release.